Reorganize pix2pix model code
