+++ examples/gan/pix2pix_tf_keras/pix2pix/generator.py
@@ -1,0 +1,78 @@
+import tensorflow as tf
+
+from .sampling import downsample, upsample
+
+
+OUTPUT_CHANNELS = 3
+
+
+def Generator():
+    inputs = tf.keras.layers.Input(shape=[256, 256, 3])
+
+    down_stack = [
+        downsample(64, 4, apply_batchnorm=False),  # (batch_size, 128, 128, 64)
+        downsample(128, 4),  # (batch_size, 64, 64, 128)
+        downsample(256, 4),  # (batch_size, 32, 32, 256)
+        downsample(512, 4),  # (batch_size, 16, 16, 512)
+        downsample(512, 4),  # (batch_size, 8, 8, 512)
+        downsample(512, 4),  # (batch_size, 4, 4, 512)
+        downsample(512, 4),  # (batch_size, 2, 2, 512)
+        downsample(512, 4),  # (batch_size, 1, 1, 512)
+    ]
+
+    up_stack = [
+        upsample(512, 4, apply_dropout=True),  # (batch_size, 2, 2, 1024)
+        upsample(512, 4, apply_dropout=True),  # (batch_size, 4, 4, 1024)
+        upsample(512, 4, apply_dropout=True),  # (batch_size, 8, 8, 1024)
+        upsample(512, 4),  # (batch_size, 16, 16, 1024)
+        upsample(256, 4),  # (batch_size, 32, 32, 512)
+        upsample(128, 4),  # (batch_size, 64, 64, 256)
+        upsample(64, 4),  # (batch_size, 128, 128, 128)
+    ]
+
+    initializer = tf.random_normal_initializer(0.0, 0.02)
+    last = tf.keras.layers.Conv2DTranspose(
+        OUTPUT_CHANNELS,
+        4,
+        strides=2,
+        padding="same",
+        kernel_initializer=initializer,
+        activation="tanh",
+    )  # (batch_size, 256, 256, 3)
+
+    x = inputs
+
+    # Downsampling through the model
+    skips = []
+    for down in down_stack:
+        x = down(x)
+        skips.append(x)
+
+    skips = reversed(skips[:-1])
+
+    # Upsampling and establishing the skip connections
+    for up, skip in zip(up_stack, skips):
+        x = up(x)
+        x = tf.keras.layers.Concatenate()([x, skip])
+
+    x = last(x)
+
+    return tf.keras.Model(inputs=inputs, outputs=x)
+
+
+LAMBDA = 100
+
+
+def loss(disc_generated_output, gen_output, target):
+    cross_entropy = tf.keras.losses.BinaryCrossentropy(from_logits=True)
+    gan_loss = cross_entropy(tf.ones_like(disc_generated_output), disc_generated_output)
+
+    # Mean absolute error
+    l1_loss = tf.reduce_mean(tf.abs(target - gen_output))
+
+    total_gen_loss = gan_loss + (LAMBDA * l1_loss)
+
+    return total_gen_loss, gan_loss, l1_loss
+
+
+optimizer = tf.keras.optimizers.Adam(2e-4, beta_1=0.5)
