+++ examples/gan/pix2pix_tf_keras/pix2pix/sampling.py
@@ -1,0 +1,49 @@
+import tensorflow as tf
+
+
+def downsample(filters, size, apply_batchnorm=True):
+    initializer = tf.random_normal_initializer(0.0, 0.02)
+
+    result = tf.keras.Sequential()
+    result.add(
+        tf.keras.layers.Conv2D(
+            filters,
+            size,
+            strides=2,
+            padding="same",
+            kernel_initializer=initializer,
+            use_bias=False,
+        )
+    )
+
+    if apply_batchnorm:
+        result.add(tf.keras.layers.BatchNormalization())
+
+    result.add(tf.keras.layers.LeakyReLU())
+
+    return result
+
+
+def upsample(filters, size, apply_dropout=False):
+    initializer = tf.random_normal_initializer(0.0, 0.02)
+
+    result = tf.keras.Sequential()
+    result.add(
+        tf.keras.layers.Conv2DTranspose(
+            filters,
+            size,
+            strides=2,
+            padding="same",
+            kernel_initializer=initializer,
+            use_bias=False,
+        )
+    )
+
+    result.add(tf.keras.layers.BatchNormalization())
+
+    if apply_dropout:
+        result.add(tf.keras.layers.Dropout(0.5))
+
+    result.add(tf.keras.layers.ReLU())
+
+    return result
